// fsd/entity/Size/model/repo/responseError.enum.ts
export enum SizeResponseErrorEnum {
  // SIZE_EXIST = "Store already existing",
  SIZE_NOT_EXIST = "Size has not been exist",
  SIZE_NOT_UNIQUE = "Size has not been unique name",
  SIZE_NOT_FOUND = "Size has not been found",
  SIZE_NOT_CREATED = "Size has not been created",
  SIZE_NOT_UPDATED = "Size has not been updated",
  SIZE_NOT_REMOVED = "Size has not been removed",
  SIZE_NO_OWNER = "You are not owner this size",
  //
  // SIZE_SLUG_EMPTY = "Slug not must be empty",
  // SIZE_LIST_NOT_FOUND = "Store list has not been found",
  // USER_NOT_FOUND = "User not found",
}